Set debugmode to true on staging-server
Because the robots-meta-tag is only inserted when debug is true, we need debug
to be true on the staging-server.

We can't solve this with ENV-variables because they aren't passed due a
documented issue with suExec and mod_env, see http://wiki.dreamhost.com/Suexec#Apache_module_mod_env
for more information.