Make sure the robots-meta-tag is added on the staging-server
Because the robots-meta-tag is only inserted when debug is true, and debug is
based on the ENV-variables we have to add an extra check.

We can't solve this with ENV-variables because they aren't passed due a
documented issue with suExec and mod_env, see http://wiki.dreamhost.com/Suexec#Apache_module_mod_env
for more information.